requirements.txt
# Core dependencies
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
jsonschema>=4.17.0

# Document processing with intelligent extraction
pdfplumber>=0.9.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
contextgem>=0.1.0

# LLM interface
litellm>=1.35.0

# Token counting
tiktoken>=0.5.0

# Simple vector similarity (optional, lightweight)
numpy>=1.24.0
scikit-learn>=1.3.0

psycopg2-binary==2.9.9
pgvector==0.2.0
